desktop: refactor cmds to be in one file
nfci
